test(config): improved the assertion and parsing of config
