+++ .github/workflows/ci.yaml
@@ -30,8 +30,6 @@ jobs:
                 node: ["18", "17", "16", "14"]
                 os: [windows-latest, ubuntu-latest]
         name: Test with Node v${{ matrix.node }} on ${{ matrix.os }}
-        env:
-            IS_CI: true
         steps:
             - name: Tell if project is using npm or yarn
               id: step1
