+++ package.json
@@ -15,7 +15,7 @@
         "postbuild": "yarn grant-exec-perms",
         "pretest": "yarn build",
         "test": "vitest",
-        "prettier": "prettier '**/*.{ts,tsx,json,md}'",
+        "prettier": "prettier --ignore-path .gitignore '**/*.{ts,tsx,json,md}'",
         "format": "yarn prettier --write",
         "format:check": "yarn prettier --list-different",
         "link_in_test_app": "node dist/bin/link_in_test_app.js"
@@ -71,6 +71,7 @@
         "@types/node": "^10.0.0",
         "@types/node-fetch": "^2.5.6",
         "@types/url-join": "^4.0.0",
+        "cases-of-test": "^0.0.0",
         "husky": "^4.3.8",
         "js-yaml": "^3.13.1",
         "lint-staged": "^11.0.0",
