+++ test/index.ts
@@ -1,30 +1,24 @@
-import testParseParams from "./parseParams";
 import testReplacer from "./replacer";
 import testGetInstalledVersionPackageJson from "./getInstalledVersionPackageJson";
 import testParsedImportExportStatementTypes from "./types/parsedImportExportStatement";
-import { Tests } from "./shared";
 import testGetValidImportUrl from "./resolveNodeModuleToDenoModule/getValidImportUrl";
 import testParseAndReadConfigFile from "./configFile";
 import testResolveNodeModuleToDenoModule from "./resolveNodeModuleToDenoModule/resolveNodeModuleToDenoModule";
 import testDenoifyImportExportStatement from "./denoifyImportExportStatement";
 import testDenoifySingleFile from "./denoifySingleFile";
+import testCases from "cases-of-test";
+import testParseParams from "./config";
 
-const tests: Tests = [
-    [testParseParams],
-    [testReplacer],
-    [testGetInstalledVersionPackageJson],
-    [testParsedImportExportStatementTypes],
-    [testGetValidImportUrl],
-    [testParseAndReadConfigFile],
-    [testResolveNodeModuleToDenoModule],
-    [testDenoifyImportExportStatement],
-    [testDenoifySingleFile]
-];
-
-const selectedTests = tests.filter(([_, only]) => only);
-
-if (process.env.IS_CI && selectedTests.length) {
-    throw new Error('cannot have "only" in ci cd');
-}
-
-(!selectedTests.length ? tests : selectedTests).forEach(([test]) => test());
+testCases({
+    tests: [
+        [testParseParams, "only"],
+        [testReplacer],
+        [testGetInstalledVersionPackageJson],
+        [testParsedImportExportStatementTypes],
+        [testGetValidImportUrl],
+        [testParseAndReadConfigFile],
+        [testResolveNodeModuleToDenoModule],
+        [testDenoifyImportExportStatement],
+        [testDenoifySingleFile]
+    ]
+});
