--- test/shared/index.ts
@@ -1,3 +1,0 @@
-type Tests = ReadonlyArray<readonly [() => void, "only"?]>;
-
-export { Tests };
